Small optimizations.

+++ src/network/netdev/vlan.h
@@ -9,7 +9,7 @@ struct VLan {
         NetDev meta;
 
         uint16_t id;
-        uint16_t protocol;
+        int protocol;
 
         int gvrp;
         int mvrp;
